// game-service/src/modules/game/core/game.session.d.ts
import { Server } from 'socket.io';
import { World } from './game.world';
import { Vector } from '../utils/game.vector';
import { GameRules } from './game.rules';
import { Engine } from './game.engine';
import { IGameState } from '../gameStates/game.state.interface';
import { PlayerManager } from '../managers/playerManager/player.manager';
import { BulletManager } from '../managers/bullet.manager';
import { Player, AttackType, MatchMakingData, MatchMode, MatchType } from "@transcendence/types";
import { GameService } from '../game.service';
import { ExitStatus } from '../interfaces-enums/exitStatus.interface';
export declare class GameSession {
    readonly gameId: string;
    readonly server: Server;
    readonly gameWorld: World;
    readonly gameRules: GameRules;
    private readonly playerManager;
    private readonly bulletManager;
    readonly gameService: GameService;
    private logger;
    readonly expectedUserDbIds: string[];
    readonly players: Map<string, Player>;
    readonly socketToEntities: Map<string, string[]>;
    readonly engine: Engine;
    private currentState;
    private sessionTime;
    matchType: MatchType;
    matchMode: MatchMode;
    constructor(gameId: string, server: Server, gameWorld: World, gameRules: GameRules, playerManager: PlayerManager, bulletManager: BulletManager, matchType: MatchType, matchMode: MatchMode, gameService: GameService);
    addPlayer(player: MatchMakingData, socketId: string | undefined): ExitStatus;
    addBot(player: MatchMakingData): ExitStatus;
    processInput(socketId: string, input: Vector, attackType: AttackType, playerIndex: number): void;
    update(dt: number): void;
    cleanUp(): void;
    isGameOver(): boolean;
    isPlaying(): boolean;
    isJoinable(): boolean;
    canShutdown(): boolean;
    removePlayer(entityId: string): void;
    sendMessage(author: Player, message: string): void;
    transitionTo(newState: IGameState): void;
    getGameState(): string;
    getGameId(): string;
    getPlayersIds(): string[];
    getPlayerIndex(): number;
}
//# sourceMappingURL=game.session.d.ts.map